Moved WeakList into utils
