+++ src/amulet/utils/weak.hpp
@@ -1,0 +1,34 @@
+#pragma once
+
+#include <list>
+#include <memory>
+
+namespace Amulet {
+
+namespace detail {
+    template <typename C, typename T>
+    void for_each(C& l, std::function<void(T&)> f)
+    {
+        auto it = l.begin();
+        while (it != l.end()) {
+            auto ptr = it->lock();
+            if (ptr) {
+                f(*ptr);
+                it++;
+            } else {
+                it = l.erase(it);
+            }
+        }
+    }
+}
+
+template <typename T>
+using WeakList = std::list<std::weak_ptr<T>>;
+
+template <typename T>
+void for_each(WeakList<T>& l, std::function<void(T&)> f)
+{
+    detail::for_each(l, f);
+}
+
+} // namespace Amulet
